Fixed bug with NamedTemporaryFile on Windows
Changed NamedTemporaryFile to be manually deleted isntead of automatic. We delete it manually because otherwise on Windows it gets deleted before we the temp file is saved to the output file location. This is because on Windows, file handles with the O_TEMPORARY flag (which is set if we pass `delete=True`) are deleted as soon as they're closed.